add all functions to populate registry
