+++ psycop/common/model_training_v2/config/static_registry_configs/data/minimal_test_data-20231115_104208.yaml
@@ -1,0 +1,2 @@
+[placeholder]
+@data = "minimal_test_data"
