+++ psycop/common/model_training_v2/config/static_registry_configs/metrics/binary_auroc-20231115_093403.yaml
@@ -1,0 +1,2 @@
+[placeholder]
+@metrics = "binary_auroc"
