+++ psycop/common/model_training_v2/config/static_registry_configs/preprocessing/age_filter-20231115_100859.yaml
@@ -1,0 +1,5 @@
+[placeholder]
+@preprocessing = "age_filter"
+min_age = 0
+max_age = 150
+age_col_name = "age"
