+++ psycop/common/model_training_v2/trainer/task/estimator_steps/logistic_regression.py
@@ -55,8 +55,8 @@ def logistic_regression_step(
 class LogisticRegressionSuggester(Suggester):
     def __init__(
         self,
-        C: FloatSpaceT,
-        l1_ratio: FloatSpaceT,
+        C: FloatSpaceT = {"low": 1e-8, "high": 10, "logarithmic": False},  # noqa: B006
+        l1_ratio: FloatSpaceT = {"low": 0.0, "high": 1, "logarithmic": False},  # noqa: B006
         solvers: Sequence[LogRegSolvers] = ("saga",),
         penalties: Sequence[LogRegPenalties] = ("l1", "l2", "elasticnet"),
     ):
